Tidier version of the MedInfo panel
